src/pyDjango/requirements.txt
absl-py==0.8.0
astor==0.8.0
astroid==2.2.5
audioread==2.1.8
autopep8==1.4.4
beautifulsoup4==4.8.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cognitive-face==1.5.0
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.0
Django==2.2.5
django-livereload-server==0.3.2
django-rest-swagger==2.2.0
djangorestframework==3.10.3
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.1
joblib==0.13.2
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.2
librosa==0.7.0
llvmlite==0.29.0
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
numba==0.45.1
numpy==1.17.1
openapi-codec==1.3.2
optional-django==0.1.0
pandas==0.25.1
protobuf==3.9.1
pycodestyle==2.5.0
pycparser==2.19
pyjsparser==2.7.1
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
resampy==0.2.2
scikit-learn==0.21.3
scipy==1.3.1
simplejson==3.16.0
six==1.12.0
SoundFile==0.10.2
soupsieve==1.9.3
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tornado==6.0.3
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.15.6
wrapt==1.11.2
